Split block-preview pages (#353)
Simplifies adding new block-preview components
