+++ site/src/app/block-preview/layout.tsx
@@ -1,9 +1,10 @@
+export const dynamic = "force-dynamic";
+
 import { GlobalStyle } from "@src/layout/GlobalStyle";
+import { SiteConfigsProvider } from "@src/util/blockPreview";
 import { ResponsiveSpacingStyle } from "@src/util/ResponsiveSpacingStyle";
 import { getSiteConfigs } from "@src/util/siteConfig";
 import StyledComponentsRegistry from "@src/util/StyledComponentsRegistry";
-
-import { SiteConfigsProvider } from "./[type]/page";
 
 export default async function BlockPreviewLayout({ children }: { children: React.ReactNode }) {
     const siteConfigs = getSiteConfigs();
