+++ site/src/app/block-preview/stage/page.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import { useIFrameBridge } from "@comet/cms-site";
+import { StageBlock } from "@src/documents/pages/blocks/StageBlock";
+import { withBlockPreview } from "@src/util/blockPreview";
+
+export default withBlockPreview(() => {
+    const { block: stage } = useIFrameBridge();
+    return stage ? <StageBlock data={stage} /> : null;
+});
